fix(adapter-utils): address sandbox git review feedback
