+++ packages/adapter-utils/src/sandbox-managed-runtime.ts
@@ -21,6 +21,7 @@ import {
   type RuntimeProgressPhase,
   type RuntimeProgressSink,
 } from "./runtime-progress.js";
+import { isRelativePathOrDescendant, shouldExcludePath } from "./exclude-patterns.js";
 
 const execFile = promisify(execFileCallback);
 const SANDBOX_WORKSPACE_HEAVY_DIR_NAMES = [
@@ -239,35 +240,6 @@ async function walkDirectory(root: string, relative = ""): Promise<string[]> {
     }
   }
   return out.sort((left, right) => right.length - left.length);
-}
-
-function isRelativePathOrDescendant(relative: string, candidate: string): boolean {
-  return relative === candidate || relative.startsWith(`${candidate}/`);
-}
-
-function pathContainsSegmentOrDescendant(relative: string, segment: string): boolean {
-  return relative === segment ||
-    relative.startsWith(`${segment}/`) ||
-    relative.endsWith(`/${segment}`) ||
-    relative.includes(`/${segment}/`);
-}
-
-function excludePatternMatches(relative: string, pattern: string): boolean {
-  if (pattern.startsWith("*/") && pattern.endsWith("/*")) {
-    return pathContainsSegmentOrDescendant(relative, pattern.slice(2, -2));
-  }
-  if (pattern.startsWith("*/")) {
-    return pathContainsSegmentOrDescendant(relative, pattern.slice(2));
-  }
-  if (pattern.endsWith("/*")) {
-    const base = pattern.slice(0, -2);
-    return relative.startsWith(`${base}/`);
-  }
-  return isRelativePathOrDescendant(relative, pattern);
-}
-
-function shouldExcludePath(relative: string, exclude: readonly string[]): boolean {
-  return exclude.some((entry) => excludePatternMatches(relative, entry));
 }
 
 async function copyWorkspaceEntry(sourceRoot: string, targetRoot: string, relative: string): Promise<void> {
